+++ src/main/java/tpAnual/com/EmailSender.java
@@ -3,7 +3,7 @@ package tpAnual.com;
 // Interface para mockear
 public interface EmailSender {
 	
-	public void enviarMensajePorDemora(int limite);
+	public void enviarMensajePorDemora(Long limite);
 
 	public void setMailAdministrador(String mailAdministrador);
 }
